core: resync destfolder on message move (ugly fix)
If you move a message to destfolder, at the end of MoveMessage cmd,
the message will not be in the destfolder zpush state, but on the phone,
the message is in destfolder.
Now if you delete the message from the server before you sync destfolder,
message will stay on the phone.

What we should really do is update destfolder state.

Signed-off-by: Etienne CHAMPETIER <etienne.champetier@fiducial.net>